task completed
modified folders structure
created new order details component
updated order details section
