+++ src/components/PrivateRoute.jsx
@@ -7,9 +7,6 @@ import React, { useContext } from 'react';
 const PrivateRoutes = () => {
 
     let [user, setUser] = useContext(ContextUser);
-
-
-
     return (
         user ? <Outlet /> : <Navigate to='/customers/login' />
     )
