--- src/components/admin-components/AdmUsers.jsx
@@ -1,11 +1,0 @@
-export default function AdmUsers() {
-
-
-
-
-    return (
-        <>
-            <p>Users</p>
-        </>
-    )
-}
